+++ src/Oci8/Oci8Connection.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Yajra\Oci8;
+namespace Vincent\Oci8;
 
 use Doctrine\DBAL\Connection as DoctrineConnection;
 use Doctrine\DBAL\Driver\OCI8\Driver as DoctrineDriver;
 use Illuminate\Database\Connection;
 use Illuminate\Database\Grammar;
 use PDO;
-use Yajra\Oci8\Query\Grammars\OracleGrammar as QueryGrammar;
-use Yajra\Oci8\Query\OracleBuilder as QueryBuilder;
-use Yajra\Oci8\Query\Processors\OracleProcessor as Processor;
-use Yajra\Oci8\Schema\Grammars\OracleGrammar as SchemaGrammar;
-use Yajra\Oci8\Schema\OracleBuilder as SchemaBuilder;
-use Yajra\Oci8\Schema\Sequence;
-use Yajra\Oci8\Schema\Trigger;
-use Yajra\Pdo\Oci8\Statement;
+use Vincent\Oci8\Query\Grammars\OracleGrammar as QueryGrammar;
+use Vincent\Oci8\Query\OracleBuilder as QueryBuilder;
+use Vincent\Oci8\Query\Processors\OracleProcessor as Processor;
+use Vincent\Oci8\Schema\Grammars\OracleGrammar as SchemaGrammar;
+use Vincent\Oci8\Schema\OracleBuilder as SchemaBuilder;
+use Vincent\Oci8\Schema\Sequence;
+use Vincent\Oci8\Schema\Trigger;
+use Vincent\Pdo\Oci8\Statement;
 
 class Oci8Connection extends Connection
 {
@@ -24,12 +24,12 @@ class Oci8Connection extends Connection
     protected $schema;
 
     /**
-     * @var \Yajra\Oci8\Schema\Sequence
+     * @var \Vincent\Oci8\Schema\Sequence
      */
     protected $sequence;
 
     /**
-     * @var \Yajra\Oci8\Schema\Trigger
+     * @var \Vincent\Oci8\Schema\Trigger
      */
     protected $trigger;
 
@@ -99,7 +99,7 @@ class Oci8Connection extends Connection
     /**
      * Get sequence class.
      *
-     * @return \Yajra\Oci8\Schema\Sequence
+     * @return \Vincent\Oci8\Schema\Sequence
      */
     public function getSequence()
     {
@@ -109,8 +109,8 @@ class Oci8Connection extends Connection
     /**
      * Set sequence class.
      *
-     * @param \Yajra\Oci8\Schema\Sequence $sequence
-     * @return \Yajra\Oci8\Schema\Sequence
+     * @param \Vincent\Oci8\Schema\Sequence $sequence
+     * @return \Vincent\Oci8\Schema\Sequence
      */
     public function setSequence(Sequence $sequence)
     {
@@ -120,7 +120,7 @@ class Oci8Connection extends Connection
     /**
      * Get oracle trigger class.
      *
-     * @return \Yajra\Oci8\Schema\Trigger
+     * @return \Vincent\Oci8\Schema\Trigger
      */
     public function getTrigger()
     {
@@ -130,8 +130,8 @@ class Oci8Connection extends Connection
     /**
      * Set oracle trigger class.
      *
-     * @param \Yajra\Oci8\Schema\Trigger $trigger
-     * @return \Yajra\Oci8\Schema\Trigger
+     * @param \Vincent\Oci8\Schema\Trigger $trigger
+     * @return \Vincent\Oci8\Schema\Trigger
      */
     public function setTrigger(Trigger $trigger)
     {
@@ -141,7 +141,7 @@ class Oci8Connection extends Connection
     /**
      * Get a schema builder instance for the connection.
      *
-     * @return \Yajra\Oci8\Schema\OracleBuilder
+     * @return \Vincent\Oci8\Schema\OracleBuilder
      */
     public function getSchemaBuilder()
     {
@@ -156,7 +156,7 @@ class Oci8Connection extends Connection
      * Begin a fluent query against a database table.
      *
      * @param string $table
-     * @return \Yajra\Oci8\Query\OracleBuilder
+     * @return \Vincent\Oci8\Query\OracleBuilder
      */
     public function table($table)
     {
@@ -196,7 +196,7 @@ class Oci8Connection extends Connection
                 $data, $this->getDoctrineDriver()
             );
         }
-        
+
         return $this->doctrineConnection;
     }
 
@@ -296,7 +296,7 @@ class Oci8Connection extends Connection
     /**
      * Get the default query grammar instance.
      *
-     * @return \Illuminate\Database\Grammar|\Yajra\Oci8\Query\Grammars\OracleGrammar
+     * @return \Illuminate\Database\Grammar|\Vincent\Oci8\Query\Grammars\OracleGrammar
      */
     protected function getDefaultQueryGrammar()
     {
@@ -306,7 +306,7 @@ class Oci8Connection extends Connection
     /**
      * Set the table prefix and return the grammar.
      *
-     * @param \Illuminate\Database\Grammar|\Yajra\Oci8\Query\Grammars\OracleGrammar|\Yajra\Oci8\Schema\Grammars\OracleGrammar $grammar
+     * @param \Illuminate\Database\Grammar|\Vincent\Oci8\Query\Grammars\OracleGrammar|\Vincent\Oci8\Schema\Grammars\OracleGrammar $grammar
      * @return \Illuminate\Database\Grammar
      */
     public function withTablePrefix(Grammar $grammar)
@@ -317,7 +317,7 @@ class Oci8Connection extends Connection
     /**
      * Set the schema prefix and return the grammar.
      *
-     * @param \Illuminate\Database\Grammar|\Yajra\Oci8\Query\Grammars\OracleGrammar|\Yajra\Oci8\Schema\Grammars\OracleGrammar $grammar
+     * @param \Illuminate\Database\Grammar|\Vincent\Oci8\Query\Grammars\OracleGrammar|\Vincent\Oci8\Schema\Grammars\OracleGrammar $grammar
      * @return \Illuminate\Database\Grammar
      */
     public function withSchemaPrefix(Grammar $grammar)
@@ -340,7 +340,7 @@ class Oci8Connection extends Connection
     /**
      * Get the default schema grammar instance.
      *
-     * @return \Illuminate\Database\Grammar|\Yajra\Oci8\Schema\Grammars\OracleGrammar
+     * @return \Illuminate\Database\Grammar|\Vincent\Oci8\Schema\Grammars\OracleGrammar
      */
     protected function getDefaultSchemaGrammar()
     {
@@ -350,7 +350,7 @@ class Oci8Connection extends Connection
     /**
      * Get the default post processor instance.
      *
-     * @return \Yajra\Oci8\Query\Processors\OracleProcessor
+     * @return \Vincent\Oci8\Query\Processors\OracleProcessor
      */
     protected function getDefaultPostProcessor()
     {