+++ src/Oci8/Query/Grammars/OracleGrammar.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Yajra\Oci8\Query\Grammars;
+namespace Vincent\Oci8\Query\Grammars;
 
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Database\Query\Grammars\Grammar;
 use Illuminate\Support\Str;
-use Yajra\Oci8\OracleReservedWords;
+use Vincent\Oci8\OracleReservedWords;
 
 class OracleGrammar extends Grammar
 {